requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
aiofiles>=24.1.0
ujson>=5.10.0
